lint: fix remaining 15 ruff errors — all checks pass
- Add noqa: N806 to observability.py dynamic class imports (intentional pattern)
- SIM108: simplify if/else to ternary in feed/service.py
- B905: add strict=False to zip() in fitness/service.py
- RUF046: remove redundant int() wrapping round() in gamification+study
- N812: replace 'date as DateType' with 'date as date_t' (lowercase alias)
- B904: add 'from exc' to raise in worker.py retry
- B017: add noqa to blind Exception in test_security.py
- Update pyproject.toml: target-version py311, ignore B008/RUF012/TC* (Pydantic patterns)
- ruff check + ruff format now both pass with 0 errors

+++ backend/app/modules/gamification/service.py
@@ -392,7 +392,7 @@ class GamificationService:
             return 5
         if action == "study_session_completed":
             duration = int(metadata.get("duration_minutes") or 0)
-            return max(0, int(round((duration / 60.0) * 20)))
+            return max(0, round((duration / 60.0) * 20))
         if action == "sleep_goal_met":
             return 25
         if action == "workout_logged":
